requirements.txt
flask<3
werkzeug<3
opentelemetry-instrumentation
opentelemetry-exporter-otlp
